Make the gem a backport of MRI 3.1 `GC.total_time` and `GC.stat(:time)`
We also go back to using tracepoint hooks. While they disable the allocation
fast path, in the end the overhead isn't measurably bigger than `GC::Profiler`
and more importantly it doesn't keep a record of the events, so we don't
have to fear about memory leaks.

The gem is a total noop on Ruby 3.1+, and a very thin backport on
JRuby and TruffleRuby.